ELO as a float

+++ chessbolt-api/models/dtos/player.py
@@ -4,4 +4,4 @@ from pydantic import BaseModel
 class PlayerDto(BaseModel):
     id: int
     name: str
-    elo: int
+    elo: float
